windows: Use WinSSL/schannel instead of OpenSSL
The prebuilt libcurl binaries use MultiSSL, which allows for
choosing the backend at runtime. This doesn't yet remove the
requirement for libssl and libcrypto, but at least we don't
need a seperate certificate store anymore.

The "load if available" snipped had to be adjusted since
GMod ignores the return code of the loaded module, so we
don't really have a way to let LUA know that the module can't
load properly (apart from checking afterwards if the CHTTP method
exists).

Closes: #1